completed till edit and delete

+++ database/migrations/0001_01_01_000003_create_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('super_agent_id')->nullable();
+            $table->integer('super_agent_id')->nullable();
             $table->foreignId('scheme_id')->nullable()->constrained('schemes')->nullOnDelete();
             $table->string('name')->nullable();
             $table->string('username')->unique();
